refactor monitoring app
* Add ADR to explain new app organization.
* Apply new app organization to the monitoring Django app.
* Remove middleware ordering checks. This is not a typical Django
feature and it is painful when refactoring.
* Add CachedCustomMonitoringMiddleware, CodeOwnerMonitoringMiddleware,
and MonitoringMemoryMiddleware to the public API.
* Deprecate above middleware from old location.
* Deprecate methods from modules that were always meant
to be used from the public API:
  * accumulate
  * increment
  * set_custom_attribute
  * set_custom_attributes_for_course_key
* Add additional custom attributes for deprecated
classes and methods to make them safer to retire.

ARCHBOM-1584
