+++ edx_django_utils/cache/middleware.py
@@ -3,8 +3,6 @@ Caching utility middleware.
 """
 from django.utils.deprecation import MiddlewareMixin
 
-from edx_django_utils.private_utils import _check_middleware_dependencies
-
 from . import RequestCache, TieredCache
 
 
@@ -36,16 +34,6 @@ class TieredCacheMiddleware(MiddlewareMixin):
     """
     Middleware to store whether or not to force django cache misses.
     """
-    def __init__(self, get_response=None):
-        super(TieredCacheMiddleware, self).__init__(get_response=get_response)
-        # checks proper dependency order as well.
-        _check_middleware_dependencies(self, required_middleware=[
-            'edx_django_utils.cache.middleware.RequestCacheMiddleware',
-            # Some Authentication Middleware also needs to be in between,
-            # but don't want to hard-code that dependency.
-            'edx_django_utils.cache.middleware.TieredCacheMiddleware',
-        ])
-
     def process_request(self, request):
         """
         Stores whether or not FORCE_CACHE_MISS_PARAM was supplied in the
