+++ edx_django_utils/cache/tests/test_middleware.py
@@ -2,7 +2,7 @@
 """
 Tests for the RequestCacheMiddleware.
 """
-from django.test import RequestFactory, TestCase, override_settings
+from django.test import RequestFactory, TestCase
 from mock import MagicMock
 
 from edx_django_utils.cache import middleware
@@ -15,7 +15,6 @@ TEST_NAMESPACE = "test_namespace"
 
 class TestRequestCacheMiddleware(TestCase):  # pylint: disable=missing-class-docstring
 
-    @override_settings(MIDDLEWARE=['edx_django_utils.cache.middleware.RequestCacheMiddleware'])
     def setUp(self):
         super(TestRequestCacheMiddleware, self).setUp()
         self.middleware = middleware.RequestCacheMiddleware()
@@ -56,10 +55,6 @@ class TestRequestCacheMiddleware(TestCase):  # pylint: disable=missing-class-doc
 
 class TestTieredCacheMiddleware(TestCase):  # pylint: disable=missing-class-docstring
 
-    @override_settings(MIDDLEWARE=[
-        'edx_django_utils.cache.middleware.RequestCacheMiddleware',
-        'edx_django_utils.cache.middleware.TieredCacheMiddleware'
-    ])
     def setUp(self):
         super(TestTieredCacheMiddleware, self).setUp()
         self.middleware = middleware.TieredCacheMiddleware()
@@ -89,11 +84,6 @@ class TestTieredCacheMiddleware(TestCase):  # pylint: disable=missing-class-docs
         self.middleware.process_request(request)
 
         self.assertFalse(self.request_cache.get_cached_response(SHOULD_FORCE_CACHE_MISS_KEY).value)
-
-    @override_settings(MIDDLEWARE=['some.Middleware'])
-    def test_tiered_cache_missing_middleware(self):
-        with self.assertRaises(AssertionError):
-            middleware.TieredCacheMiddleware()
 
     def _mock_user(self, is_staff=True):
         mock_user = MagicMock()
