+++ edx_django_utils/monitoring/__init__.py
@@ -3,14 +3,15 @@ Metrics utilities public api
 
 See README.rst for details.
 """
-from .code_owner.utils import get_code_owner_from_module
-from .transactions import function_trace, get_current_transaction, ignore_transaction, set_monitoring_transaction_name
-# "set_custom_metric*" methods are deprecated
-from .utils import (
-    accumulate,
-    increment,
-    set_custom_attribute,
-    set_custom_attributes_for_course_key,
-    set_custom_metric,
-    set_custom_metrics_for_course_key
+from .internal.code_owner.middleware import CodeOwnerMonitoringMiddleware
+from .internal.code_owner.utils import get_code_owner_from_module
+from .internal.middleware import CachedCustomMonitoringMiddleware, MonitoringMemoryMiddleware
+from .internal.transactions import (
+    function_trace,
+    get_current_transaction,
+    ignore_transaction,
+    set_monitoring_transaction_name
 )
+from .internal.utils import accumulate, increment, set_custom_attribute, set_custom_attributes_for_course_key
+# "set_custom_metric*" methods are deprecated
+from .utils import set_custom_metric, set_custom_metrics_for_course_key
