+++ edx_django_utils/monitoring/tests/code_owner/test_middleware.py
@@ -10,9 +10,10 @@ from django.test import RequestFactory, override_settings
 from django.views.generic import View
 from mock import Mock, call, patch
 
-from edx_django_utils.monitoring.code_owner.middleware import CodeOwnerMonitoringMiddleware
-from edx_django_utils.monitoring.code_owner.tests.mock_views import MockViewTest
-from edx_django_utils.monitoring.code_owner.utils import clear_cached_mappings
+from edx_django_utils.monitoring import CodeOwnerMonitoringMiddleware
+from edx_django_utils.monitoring.internal.code_owner.utils import clear_cached_mappings
+
+from .mock_views import MockViewTest
 
 
 class MockMiddlewareViewTest(View):
@@ -47,15 +48,15 @@ class CodeOwnerMetricMiddlewareTests(TestCase):
         self.assertEqual(self.middleware(request), 'test-response')
 
     _REQUEST_PATH_TO_MODULE_PATH = {
-        '/middleware-test/': 'edx_django_utils.monitoring.code_owner.tests.test_middleware',
-        '/test/': 'edx_django_utils.monitoring.code_owner.tests.mock_views',
+        '/middleware-test/': 'edx_django_utils.monitoring.tests.code_owner.test_middleware',
+        '/test/': 'edx_django_utils.monitoring.tests.code_owner.mock_views',
     }
 
     @override_settings(
-        CODE_OWNER_MAPPINGS={'team-red': ['edx_django_utils.monitoring.code_owner.tests.mock_views']},
+        CODE_OWNER_MAPPINGS={'team-red': ['edx_django_utils.monitoring.tests.code_owner.mock_views']},
         ROOT_URLCONF=__name__,
     )
-    @patch('edx_django_utils.monitoring.code_owner.middleware.set_custom_attribute')
+    @patch('edx_django_utils.monitoring.internal.code_owner.middleware.set_custom_attribute')
     @ddt.data(
         ('/middleware-test/', None),
         ('/test/', 'team-red'),
@@ -79,12 +80,12 @@ class CodeOwnerMetricMiddlewareTests(TestCase):
 
     @override_settings(
         CODE_OWNER_MAPPINGS={
-            'team-red': ['edx_django_utils.monitoring.code_owner.tests.mock_views'],
+            'team-red': ['edx_django_utils.monitoring.tests.code_owner.mock_views'],
             'team-blue': ['*'],
         },
         ROOT_URLCONF=__name__,
     )
-    @patch('edx_django_utils.monitoring.code_owner.middleware.set_custom_attribute')
+    @patch('edx_django_utils.monitoring.internal.code_owner.middleware.set_custom_attribute')
     @ddt.data(
         ('/middleware-test/', 'team-blue'),
         ('/test/', 'team-red'),
@@ -101,20 +102,20 @@ class CodeOwnerMetricMiddlewareTests(TestCase):
         )
 
     @override_settings(
-        CODE_OWNER_MAPPINGS={'team-red': ['edx_django_utils.monitoring.code_owner.tests.mock_views']},
+        CODE_OWNER_MAPPINGS={'team-red': ['edx_django_utils.monitoring.tests.code_owner.mock_views']},
         ROOT_URLCONF=__name__,
     )
-    @patch('edx_django_utils.monitoring.code_owner.middleware.set_custom_attribute')
+    @patch('edx_django_utils.monitoring.internal.code_owner.middleware.set_custom_attribute')
     @patch('newrelic.agent')
     @ddt.data(
         (
-            'edx_django_utils.monitoring.code_owner.tests.test_middleware',
-            'edx_django_utils.monitoring.code_owner.tests.test_middleware:MockMiddlewareViewTest',
+            'edx_django_utils.monitoring.tests.code_owner.test_middleware',
+            'edx_django_utils.monitoring.tests.code_owner.test_middleware:MockMiddlewareViewTest',
             None
         ),
         (
-            'edx_django_utils.monitoring.code_owner.tests.mock_views',
-            'edx_django_utils.monitoring.code_owner.tests.mock_views:MockViewTest',
+            'edx_django_utils.monitoring.tests.code_owner.mock_views',
+            'edx_django_utils.monitoring.tests.code_owner.mock_views:MockViewTest',
             'team-red'
         ),
     )
@@ -139,22 +140,22 @@ class CodeOwnerMetricMiddlewareTests(TestCase):
 
     @override_settings(
         CODE_OWNER_MAPPINGS={
-            'team-red': ['edx_django_utils.monitoring.code_owner.tests.mock_views'],
+            'team-red': ['edx_django_utils.monitoring.tests.code_owner.mock_views'],
             'team-blue': ['*'],
         },
         ROOT_URLCONF=__name__,
     )
-    @patch('edx_django_utils.monitoring.code_owner.middleware.set_custom_attribute')
+    @patch('edx_django_utils.monitoring.internal.code_owner.middleware.set_custom_attribute')
     @patch('newrelic.agent')
     @ddt.data(
         (
-            'edx_django_utils.monitoring.code_owner.tests.test_middleware',
-            'edx_django_utils.monitoring.code_owner.tests.test_middleware:MockMiddlewareViewTest',
+            'edx_django_utils.monitoring.tests.code_owner.test_middleware',
+            'edx_django_utils.monitoring.tests.code_owner.test_middleware:MockMiddlewareViewTest',
             'team-blue'
         ),
         (
-            'edx_django_utils.monitoring.code_owner.tests.mock_views',
-            'edx_django_utils.monitoring.code_owner.tests.mock_views:MockViewTest',
+            'edx_django_utils.monitoring.tests.code_owner.mock_views',
+            'edx_django_utils.monitoring.tests.code_owner.mock_views:MockViewTest',
             'team-red'
         ),
     )
@@ -171,10 +172,10 @@ class CodeOwnerMetricMiddlewareTests(TestCase):
         )
 
     @override_settings(
-        CODE_OWNER_MAPPINGS={'team-red': ['edx_django_utils.monitoring.code_owner.tests.mock_views']},
+        CODE_OWNER_MAPPINGS={'team-red': ['edx_django_utils.monitoring.tests.code_owner.mock_views']},
         ROOT_URLCONF=__name__,
     )
-    @patch('edx_django_utils.monitoring.code_owner.middleware.set_custom_attribute')
+    @patch('edx_django_utils.monitoring.internal.code_owner.middleware.set_custom_attribute')
     @patch('newrelic.agent')
     def test_code_owner_transaction_mapping_error(self, mock_newrelic_agent, mock_set_custom_attribute):
         mock_newrelic_agent.current_transaction = Mock(side_effect=Exception('forced exception'))
@@ -184,13 +185,13 @@ class CodeOwnerMetricMiddlewareTests(TestCase):
             mock_set_custom_attribute, has_path_error=True, has_transaction_error=True
         )
 
-    @patch('edx_django_utils.monitoring.code_owner.middleware.set_custom_attribute')
+    @patch('edx_django_utils.monitoring.internal.code_owner.middleware.set_custom_attribute')
     def test_code_owner_no_mappings(self, mock_set_custom_attribute):
         request = RequestFactory().get('/test/')
         self.middleware(request)
         mock_set_custom_attribute.assert_not_called()
 
-    @patch('edx_django_utils.monitoring.code_owner.middleware.set_custom_attribute')
+    @patch('edx_django_utils.monitoring.internal.code_owner.middleware.set_custom_attribute')
     def test_code_owner_transaction_no_mappings(self, mock_set_custom_attribute):
         request = RequestFactory().get('/bad/path/')
         self.middleware(request)
@@ -199,7 +200,7 @@ class CodeOwnerMetricMiddlewareTests(TestCase):
     @override_settings(
         CODE_OWNER_MAPPINGS={'team-red': ['lms.djangoapps.monitoring.tests.mock_views']},
     )
-    @patch('edx_django_utils.monitoring.code_owner.middleware.set_custom_attribute')
+    @patch('edx_django_utils.monitoring.internal.code_owner.middleware.set_custom_attribute')
     def test_no_resolver_for_path_and_no_transaction(self, mock_set_custom_attribute):
         request = RequestFactory().get('/bad/path/')
         self.middleware(request)
@@ -210,7 +211,7 @@ class CodeOwnerMetricMiddlewareTests(TestCase):
     @override_settings(
         CODE_OWNER_MAPPINGS={'team-red': ['*']},
     )
-    @patch('edx_django_utils.monitoring.code_owner.middleware.set_custom_attribute')
+    @patch('edx_django_utils.monitoring.internal.code_owner.middleware.set_custom_attribute')
     def test_catch_all_instead_of_errors(self, mock_set_custom_attribute):
         request = RequestFactory().get('/bad/path/')
         self.middleware(request)
@@ -220,7 +221,7 @@ class CodeOwnerMetricMiddlewareTests(TestCase):
         CODE_OWNER_MAPPINGS=['invalid_setting_as_list'],
         ROOT_URLCONF=__name__,
     )
-    @patch('edx_django_utils.monitoring.code_owner.middleware.set_custom_attribute')
+    @patch('edx_django_utils.monitoring.internal.code_owner.middleware.set_custom_attribute')
     def test_load_config_with_invalid_dict(self, mock_set_custom_attribute):
         request = RequestFactory().get('/test/')
         self.middleware(request)
