+++ edx_django_utils/monitoring/tests/test_monitoring.py
@@ -1,18 +1,22 @@
 """
 Tests for custom monitoring.
 """
-from django.test import TestCase, override_settings
+import ddt
+from django.test import TestCase
 from mock import call, patch
 
-from edx_django_utils import monitoring
 from edx_django_utils.cache import RequestCache
-from edx_django_utils.monitoring.middleware import (
-    CachedCustomMonitoringMiddleware,
-    MonitoringCustomMetricsMiddleware,
-    MonitoringMemoryMiddleware
-)
+from edx_django_utils.monitoring import CachedCustomMonitoringMiddleware, accumulate, get_current_transaction, increment
+
+from ..middleware import CachedCustomMonitoringMiddleware as DeprecatedCachedCustomMonitoringMiddleware
+from ..middleware import MonitoringCustomMetricsMiddleware as DeprecatedMonitoringCustomMetricsMiddleware
+from ..utils import accumulate as deprecated_accumulate
+from ..utils import increment as deprecated_increment
+from ..utils import set_custom_attribute as deprecated_set_custom_attribute
+from ..utils import set_custom_attributes_for_course_key as deprecated_set_custom_attributes_for_course_key
 
 
+@ddt.ddt
 class TestCustomMonitoringMiddleware(TestCase):
     """
     Test the monitoring_utils middleware and helpers
@@ -22,19 +26,24 @@ class TestCustomMonitoringMiddleware(TestCase):
         RequestCache.clear_all_namespaces()
 
     @patch('newrelic.agent')
-    @override_settings(MIDDLEWARE=[
-        'edx_django_utils.cache.middleware.RequestCacheMiddleware',
-        'edx_django_utils.monitoring.middleware.CachedCustomMonitoringMiddleware',
-    ])
-    def test_accumulate_and_increment(self, mock_newrelic_agent):
+    @ddt.data(
+        (CachedCustomMonitoringMiddleware, False, 'process_response'),
+        (CachedCustomMonitoringMiddleware, False, 'process_exception'),
+        (DeprecatedCachedCustomMonitoringMiddleware, True, 'process_response'),
+        (DeprecatedMonitoringCustomMetricsMiddleware, True, 'process_response'),
+    )
+    @ddt.unpack
+    def test_accumulate_and_increment(
+            self, cached_monitoring_middleware_class, is_deprecated, middleware_method_name, mock_newrelic_agent
+    ):
         """
         Test normal usage of collecting custom attributes and reporting to New Relic
         """
-        monitoring.accumulate('hello', 10)
-        monitoring.accumulate('world', 10)
-        monitoring.accumulate('world', 10)
-        monitoring.increment('foo')
-        monitoring.increment('foo')
+        accumulate('hello', 10)
+        accumulate('world', 10)
+        accumulate('world', 10)
+        increment('foo')
+        increment('foo')
 
         # based on the attribute data above, we expect the following calls to newrelic:
         nr_agent_calls_expected = [
@@ -44,13 +53,16 @@ class TestCustomMonitoringMiddleware(TestCase):
         ]
 
         # fake a response to trigger attributes reporting
-        CachedCustomMonitoringMiddleware().process_response(
+        middleware_method = getattr(cached_monitoring_middleware_class(), middleware_method_name)
+        middleware_method(
             'fake request',
             'fake response',
         )
 
         # Assert call counts to newrelic.agent.add_custom_parameter()
         expected_call_count = len(nr_agent_calls_expected)
+        if is_deprecated:
+            expected_call_count += 1
         measured_call_count = mock_newrelic_agent.add_custom_parameter.call_count
         self.assertEqual(expected_call_count, measured_call_count)
 
@@ -59,16 +71,20 @@ class TestCustomMonitoringMiddleware(TestCase):
         mock_newrelic_agent.add_custom_parameter.assert_has_calls(nr_agent_calls_expected, any_order=True)
 
     @patch('newrelic.agent')
-    @override_settings(MIDDLEWARE=[
-        'edx_django_utils.cache.middleware.RequestCacheMiddleware',
-        'edx_django_utils.monitoring.middleware.MonitoringCustomMetricsMiddleware',
-    ])
-    def test_accumulate_with_illegal_value(self, mock_newrelic_agent):
+    @ddt.data(
+        (CachedCustomMonitoringMiddleware, False),
+        (DeprecatedCachedCustomMonitoringMiddleware, True),
+        (DeprecatedMonitoringCustomMetricsMiddleware, True),
+    )
+    @ddt.unpack
+    def test_accumulate_with_illegal_value(
+            self, cached_monitoring_middleware_class, is_deprecated, mock_newrelic_agent
+    ):
         """
         Test monitoring accumulate with illegal value that can't be added.
         """
-        monitoring.accumulate('hello', None)
-        monitoring.accumulate('hello', 10)
+        accumulate('hello', None)
+        accumulate('hello', 10)
 
         # based on the metric data above, we expect the following calls to newrelic:
         nr_agent_calls_expected = [
@@ -77,56 +93,47 @@ class TestCustomMonitoringMiddleware(TestCase):
         ]
 
         # fake a response to trigger metrics reporting
-        MonitoringCustomMetricsMiddleware().process_response(
+        cached_monitoring_middleware_class().process_response(
             'fake request',
             'fake response',
         )
 
         # Assert call counts to newrelic.agent.add_custom_parameter()
         expected_call_count = len(nr_agent_calls_expected)
+        if is_deprecated:
+            expected_call_count += 1
         measured_call_count = mock_newrelic_agent.add_custom_parameter.call_count
         self.assertEqual(expected_call_count, measured_call_count)
 
         # Assert call args to newrelic.agent.add_custom_parameter().
         mock_newrelic_agent.add_custom_parameter.assert_has_calls(nr_agent_calls_expected, any_order=True)
 
-    @override_settings(MIDDLEWARE=[
-        'edx_django_utils.cache.middleware.RequestCacheMiddleware',
-        'edx_django_utils.monitoring.middleware.CachedCustomMonitoringMiddleware',
-    ])
-    def test_cached_custom_monitoring_middleware_dependencies_success(self):
-        CachedCustomMonitoringMiddleware()
-
-    @override_settings(MIDDLEWARE=[
-        'edx_django_utils.cache.middleware.RequestCacheMiddleware',
-        'edx_django_utils.monitoring.middleware.MonitoringCustomMetricsMiddleware',
-    ])
-    def test_custom_metrics_middleware_dependencies_success(self):
-        MonitoringCustomMetricsMiddleware()
-
-    @override_settings(MIDDLEWARE=['some.Middleware'])
-    def test_cached_custom_monitoring_middleware_dependencies_failure(self):
-        with self.assertRaises(AssertionError):
-            CachedCustomMonitoringMiddleware()
-
-    @override_settings(MIDDLEWARE=[
-        'edx_django_utils.cache.middleware.RequestCacheMiddleware',
-        'edx_django_utils.monitoring.middleware.MonitoringMemoryMiddleware',
-    ])
-    def test_memory_middleware_dependencies_success(self):
-        MonitoringMemoryMiddleware()
-
-    @override_settings(MIDDLEWARE=['some.Middleware'])
-    def test_memory_middleware_dependencies_failure(self):
-        with self.assertRaises(AssertionError):
-            MonitoringMemoryMiddleware()
-
     @patch('newrelic.agent')
     def test_get_current_transaction(self, mock_newrelic_agent):
         mock_newrelic_agent.current_transaction().name = 'fake-transaction'
-        current_transaction = monitoring.get_current_transaction()
+        current_transaction = get_current_transaction()
         self.assertEqual(current_transaction.name, 'fake-transaction')
 
     def test_get_current_transaction_without_newrelic(self):
-        current_transaction = monitoring.get_current_transaction()
+        current_transaction = get_current_transaction()
         self.assertEqual(current_transaction.name, None)
+
+    @patch('edx_django_utils.monitoring.utils.internal_accumulate')
+    def test_deprecated_accumulate(self, mock_accumulate):
+        deprecated_accumulate('key', 1)
+        mock_accumulate.assert_called_with('key', 1)
+
+    @patch('edx_django_utils.monitoring.utils.internal_increment')
+    def test_deprecated_increment(self, mock_increment):
+        deprecated_increment('key')
+        mock_increment.assert_called_with('key')
+
+    @patch('edx_django_utils.monitoring.utils.internal_set_custom_attribute')
+    def test_deprecated_set_custom_attribute(self, mock_set_custom_attribute):
+        deprecated_set_custom_attribute('key', True)
+        mock_set_custom_attribute.assert_called_with('key', True)
+
+    @patch('edx_django_utils.monitoring.utils.internal_set_custom_attributes_for_course_key')
+    def test_deprecated_set_custom_attributes_for_course_key(self, mock_set_custom_attributes_for_course_key):
+        deprecated_set_custom_attributes_for_course_key('key')
+        mock_set_custom_attributes_for_course_key.assert_called_with('key')
