+++ test_settings.py
@@ -37,6 +37,7 @@ DATABASES = {
 INSTALLED_APPS = (
     "django.contrib.auth",
     "django.contrib.contenttypes",
+    "waffle",
     "edx_django_utils",
 )
